+++ src/ops/op_enum.h
@@ -28,6 +28,8 @@ typedef enum {
     E_OP_X,
     E_OP_Y,
     E_OP_Z,
+    E_OP_J,
+    E_OP_K,
     E_OP_INIT,
     E_OP_INIT_SCENE,
     E_OP_INIT_SCRIPT,