+++ src/ops/variables.c
@@ -34,6 +34,14 @@ static void op_TIME_ACT_get(const void *data, scene_state_t *ss,
                             exec_state_t *es, command_state_t *cs);
 static void op_TIME_ACT_set(const void *data, scene_state_t *ss,
                             exec_state_t *es, command_state_t *cs);
+static void op_J_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                     command_state_t *cs);
+static void op_J_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                     command_state_t *cs);
+static void op_K_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                     command_state_t *cs);
+static void op_K_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                     command_state_t *cs);
 
 // clang-format off
 const tele_op_t op_A          = MAKE_SIMPLE_VARIABLE_OP(A         , variables.a         );
@@ -59,6 +67,8 @@ const tele_op_t op_DRUNK = MAKE_GET_SET_OP(DRUNK, op_DRUNK_get, op_DRUNK_set, 0,
 const tele_op_t op_FLIP  = MAKE_GET_SET_OP(FLIP , op_FLIP_get , op_FLIP_set , 0, true);
 const tele_op_t op_O     = MAKE_GET_SET_OP(O    , op_O_get    , op_O_set    , 0, true);
 const tele_op_t op_I     = MAKE_GET_SET_OP(I    , op_I_get, op_I_set, 0, true);
+const tele_op_t op_J     = MAKE_GET_SET_OP(J    , op_J_get, op_J_set, 0, true);
+const tele_op_t op_K     = MAKE_GET_SET_OP(K    , op_K_get, op_K_set, 0, true);
 // clang-format on
 
 static void op_TIME_get(const void *NOTUSED(data), scene_state_t *ss,
@@ -158,4 +168,28 @@ static void op_I_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
 static void op_I_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
                      exec_state_t *es, command_state_t *cs) {
     es_variables(es)->i = cs_pop(cs);
+}
+
+static void op_J_get(const void *NOTUSED(data), scene_state_t *ss,
+                     exec_state_t *es, command_state_t *cs) {
+    int16_t sn = es_variables(es)->script_number;
+    cs_push(cs, ss->variables.j[sn]);
+}
+
+static void op_J_set(const void *NOTUSED(data), scene_state_t *ss,
+                     exec_state_t *es, command_state_t *cs) {
+    int16_t sn = es_variables(es)->script_number;
+    ss->variables.j[sn] = cs_pop(cs);
+}
+
+static void op_K_get(const void *NOTUSED(data), scene_state_t *ss,
+                     exec_state_t *es, command_state_t *cs) {
+    int16_t sn = es_variables(es)->script_number;
+    cs_push(cs, ss->variables.k[sn]);
+}
+
+static void op_K_set(const void *NOTUSED(data), scene_state_t *ss,
+                     exec_state_t *es, command_state_t *cs) {
+    int16_t sn = es_variables(es)->script_number;
+    ss->variables.k[sn] = cs_pop(cs);
 }