+++ src/ops/variables.h
@@ -25,5 +25,7 @@ extern const tele_op_t op_LAST;
 extern const tele_op_t op_X;
 extern const tele_op_t op_Y;
 extern const tele_op_t op_Z;
+extern const tele_op_t op_J;
+extern const tele_op_t op_K;
 
 #endif