+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **NEW**: new op: DEL.X
 - **NEW**: new op: DEL.R
 - **IMP**: DELAY_SIZE increased to 16 from 8
+- **NEW**: new variables: J & K local script variables 
 
 ## v3.0